--- .vscode/settings.json
@@ -1,11 +1,0 @@
-{
-    "sqltools.connections": [
-        {
-            "previewLimit": 50,
-            "driver": "SQLite",
-            "name": "iCloud-Photos-Sync",
-            "database": "${workspaceFolder:icloud-photos-sync}/debug-fs/photos-library.db"
-        }
-    ],
-    "sqltools.useNodeRuntime": true
-}
