+++ rootfs/opt/icloud-photos-sync/src/lib/cli.ts
@@ -2,10 +2,10 @@ import {Command, Option, OptionValues} from 'commander';
 import chalk from 'chalk';
 import {PACKAGE_INFO} from './package.js';
 import {iCloud} from './icloud/icloud.js';
-import * as ICLOUD from './icloud/icloud.constants.js';
-import * as SYNC_ENGINE from './sync/sync-engine.constants.js';
-import * as PHOTOS_LIBRARY from './photos-library/photos-library.constants.js';
-import {SyncEngine} from './sync/sync-engine.js';
+import * as ICLOUD from './icloud/constants.js';
+import * as SYNC_ENGINE from './sync-engine/constants.js';
+import * as PHOTOS_LIBRARY from './photos-library/constants.js';
+import {SyncEngine} from './sync-engine/sync-engine.js';
 import {SingleBar} from 'cli-progress';
 import {PhotosLibrary} from './photos-library/photos-library.js';
 import {exit} from 'process';
