+++ rootfs/opt/icloud-photos-sync/src/lib/icloud/constants.ts
