+++ rootfs/opt/icloud-photos-sync/src/lib/icloud/icloud-photos/constants.ts
