+++ rootfs/opt/icloud-photos-sync/src/lib/icloud/icloud-photos/query-parser.ts
