+++ rootfs/opt/icloud-photos-sync/src/lib/logger.ts
@@ -7,7 +7,7 @@ import * as path from 'path';
 /**
  * Logger color definitions
  */
-export const colors = {
+const colors = {
     TRACE: chalk.magenta,
     DEBUG: chalk.cyan,
     INFO: chalk.blue,
