+++ rootfs/opt/icloud-photos-sync/src/lib/photos-library/constants.ts
@@ -5,11 +5,3 @@ export enum EVENTS {
     READY = `ready`,
     ERROR = `error`
 }
-
-export enum RecordState {
-    NEW = 0,
-    CHANGED = 1,
-    STALE = 2,
-    SYNCED = 3,
-    ARCHIVED = 4
-}
