+++ rootfs/opt/icloud-photos-sync/src/lib/sync-engine/constants.ts
